tests: move employee detail view to matching directory
